Project Structure and Basic Classes
Basic classes for rendering and data management.

Physics Objects can be created and rendered through the simulator (which we can have multiple of), and both their linear and angular movement can be accurately rendered.

Physics Objects (for now) can only be squares - to be changed in the future.
Force and collision manager files created, but not implemented yet.